python/requirements.txt
openai-whisper
ffmpeg-python
requests
pydub